Trying to split tests
